# Organization/Organization.Client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Organization.Server",
    "version": "1.0"
  },
  "paths": {
...
    "/api/VacationVoucher": {
      "get": {
        "tags": [
          "VacationVoucher"
        ],
        "summary": "The method returns all the vacation vouchers in the organization",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetVacationVoucherDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetVacationVoucherDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetVacationVoucherDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "VacationVoucher"
        ],
        "summary": "The method adds a new VacationVoucher into organization",
        "requestBody": {
          "description": "A new VacationVoucher that needs to be added",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostVacationVoucherDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostVacationVoucherDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostVacationVoucherDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetVacationVoucherDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetVacationVoucherDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetVacationVoucherDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/VacationVoucher/{id}": {
      "get": {
        "tags": [
          "VacationVoucher"
        ],
        "summary": "The method returns an VacationVoucher by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "VacationVoucher ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetVacationVoucherDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetVacationVoucherDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetVacationVoucherDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "VacationVoucher"
        ],
        "summary": "The method updates an VacationVoucher information by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "An ID of the VacationVoucher",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "New information of the VacationVoucher",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostVacationVoucherDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostVacationVoucherDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostVacationVoucherDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetVacationVoucherDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetVacationVoucherDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetVacationVoucherDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "VacationVoucher"
        ],
        "summary": "The method deletes an VacationVoucher by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "An ID of the VacationVoucher",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetVacationVoucherDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetVacationVoucherDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetVacationVoucherDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GetVacationVoucherDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Id - an id of a VacationVoucher",
            "format": "int32"
          },
          "issueDate": {
            "type": "string",
            "description": "IssueDate - a date, when the VacationVoucher was issued",
            "format": "date-time"
          },
          "voucherTypeId": {
            "type": "integer",
            "description": "VoucherTypeId - an id of VoucherType object",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "PostVacationVoucherDto - represents a vacation voucher, that may be issued to an employee.\r\nThe class stores information about issue date and voucher type."
      },
...
      "PostVacationVoucherDto": {
        "type": "object",
        "properties": {
          "issueDate": {
            "type": "string",
            "description": "IssueDate - a date, when the VacationVoucher was issued",
            "format": "date-time"
          },
          "voucherTypeId": {
            "type": "integer",
            "description": "VoucherTypeId - an id of VoucherType object",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "PostVacationVoucherDto - represents a vacation voucher, that may be issued to an employee.\r\nThe class stores information about issue date and voucher type."
      },
...
    }
  }
}